Parallel Dependency Analysis  (OOD, C++)
This project used parser from code analyzer project and using that I
created type table which crated a table containing all the information
about all defined types in a file, and this analysis was run on multiple
selected files using threading and each partial table by each thread was
joined to create final type table for analysis. Using this type
information dependency analysis was done and it returned which file
depends on which files or files using the type table, this analysis was
also run using thread pool and it created final dependency table showing
all the dependencies.